requirements.txt
ca-certificates==2022.10.11
certifi==2022.9.24
click==8.0.4
dataclasses==0.8
flask==2.1.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
jinja2==3.1.2
libcxx==14.0.6
libffi==3.4.2
markupsafe==2.1.1
mysql-connector-python==8.0.31
ncurses==6.3
openssl==1.1.1s
pip==22.2.2
protobuf==3.20.1
python==3.9.15
readline==8.2
setuptools==65.5.0
sqlite==3.39.3
tk==8.6.12
tzdata==2022f
werkzeug==2.0.3
wheel==0.37.1
xz==5.2.6
zipp==3.8.0
zlib==1.2.13